Refactor functions from 01 df to hf notebook into utils
